Added selenium test to mocha
